docs/backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Environment variables
python-dotenv==1.0.0

# Email validation
email-validator==2.1.0

# Testing (optional but recommended)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code quality (optional but recommended)
black==24.1.1
flake8==7.0.0
mypy==1.8.0